+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unreleased
 * Add `processor_response_type` to `Transaction`, `AuthorizationAdjustment`, and `CreditCardVerification`.
 * Add `authorizationExpiresAt` to `Transaction`
+* Fix `webhookNotification.grantedPaymentInstrumentUpdate` typo to be the correct name
 
 ## 2.13.1
 * Fix `transaction.lineItems` method when returned from transaction searches (#146)