+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `authorizationExpiresAt` to `Transaction`
 * Fix `webhookNotification.grantedPaymentInstrumentUpdate` typo to be the correct name
 * Allow PayPal payment ID and payer ID to be passed during transaction create
+* Add `fraudServiceProvider` field in `riskData`
 
 ## 2.13.1
 * Fix `transaction.lineItems` method when returned from transaction searches (#146)