+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix `webhookNotification.grantedPaymentInstrumentUpdate` typo to be the correct name
 * Allow PayPal payment ID and payer ID to be passed during transaction create
 * Add `fraudServiceProvider` field in `riskData`
+* Add `travel_flight` support to industry-specific data
 
 ## 2.13.1
 * Fix `transaction.lineItems` method when returned from transaction searches (#146)